+++ src/main/java/insane96mcp/iguanatweaksreborn/module/mobs/villager/villagers/Villagers.java
@@ -51,9 +51,6 @@ public class Villagers extends Feature {
 	@Config
 	@Label(name = "Clamp Negative Demand", description = "When villagers restock, they update the 'demand'. Demand is a trade modifier that increases the price whenever a trade is done many times, BUT when a trade is not performed, at each restock the 'demand' goes negative, making possible for a trade to never increase it's price due to high negative demand. With this to true, negative demand will be capped at -max_uses of the trade (e.g. Carrot trade from a farmer will have it's minimum demand set to -16).")
 	public static Boolean clampNegativeDemand = true;
-	@Config(min = 0)
-	@Label(name = "Heal chance", description = "1 in X chance each tick for villagers to regain 1 health. Set to 0 to disable")
-	public static Integer healChance = 200;
 	@Config(min = 0, max = 1)
 	@Label(name = "Nitwit chance", description = "Chance for a nitwit to spawn when two villagers breed")
 	public static Double nitwitChance = 0.1d;
@@ -103,7 +100,6 @@ public class Villagers extends Feature {
 			return;
 
 		tryRemovingCureDiscount(villager);
-		tryHeal(villager);
 	}
 
 	public void tryRemovingCureDiscount(Villager villager) {
@@ -117,14 +113,6 @@ public class Villagers extends Feature {
 			villager.getGossips().remove(uuid, GossipType.MINOR_POSITIVE);
 		}));
 		villager.getPersistentData().putBoolean(CURE_DISCOUNT_REMOVED, true);
-	}
-
-	public void tryHeal(Villager villager) {
-		if (healChance == 0
-				|| villager.getRandom().nextInt(healChance) != 0)
-			return;
-
-		villager.heal(1f);
 	}
 
 	public static int clampSpecialPrice(int specialPriceDiff, final ItemStack baseCostA) {